+++ src/test/shared/data/test-codesystem-versions.ts
@@ -1,0 +1,23 @@
+/*
+    Copyright (c) 2025 gematik GmbH
+    Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
+    European Commission – subsequent versions of the EUPL (the "Licence").
+    You may not use this work except in compliance with the Licence.
+    You find a copy of the Licence in the "Licence" file or at
+    https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the Licence is distributed on an "AS IS" basis,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
+    In case of changes by gematik find details in the "Readme" file.
+    See the Licence for the specific language governing permissions and limitations under the Licence.
+    *******
+    For additional notes and disclaimer from gematik and in case of changes by gematik,
+    find details in the "Readme" file.
+ */
+
+export const EXAMPLE_CODESYSTEM_VERSIONS = [
+  'http://fhir.de/CodeSystem/ifa/pei/vaccine-product|1.0.0',
+  'http://terminology.hl7.org/CodeSystem/v3-NullFlavor|2.1.0',
+  'http://snomed.info/sct|2023-03',
+  'invalid-format-without-pipe',
+];